pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["wesley1263 <weslei.andrade.souza@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
pydantic = {extras = ["email"], version = "^2.11.3"}
python-jose = "^3.4.0"
passlib = "^1.7.4"
google = "^3.0.0"
lib = "^4.0.0"
google-auth = "^2.38.0"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.166.0"
pymongo = "^4.12.0"
motor = "^3.7.0"
jwt = "^1.3.1"
python-decouple = "^3.8"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
autoflake = "^2.3.0"
pre-commit = "^3.6.2"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
